Update paps.js
chat and user panels now rezise with the window and update properly on redraw.
the chat now has a history in memory, to repopulate the chat window when its redrawn.  The chat and user panels are no longer mouseover driven for popping out, now you click them to expand and mouseout to contract. When recieving chat messages, the minimized chat panel will turn orange, when you click it to view the messages it goes back to normal.